standalone model_rotary, move model_alibi and additional variants to not_working to indicate that the code has not been adjusted/tested
